src/test/java/com/jforex/programming/position/test/PositionUtilTest.java
package com.jforex.programming.position.test;

import static org.hamcrest.Matchers.equalTo;
import static org.junit.Assert.assertThat;

import org.junit.Before;
import org.junit.Test;
import org.junit.runner.RunWith;
import org.mockito.Mock;

import com.jforex.programming.position.Position;
import com.jforex.programming.position.PositionFactory;
import com.jforex.programming.position.PositionUtil;
import com.jforex.programming.test.common.InstrumentUtilForTest;

import de.bechte.junit.runners.context.HierarchicalContextRunner;

@RunWith(HierarchicalContextRunner.class)
public class PositionUtilTest extends InstrumentUtilForTest {

    private PositionUtil positionUtil;

    @Mock
    private PositionFactory positionFactoryMock;
    @Mock
    private Position positionMock;

    @Before
    public void setUp() {
        positionUtil = new PositionUtil(positionFactoryMock);
    }

    @Test
    public void positionOrdersIsCorrect() {
        when(positionFactoryMock.forInstrument(instrumentEURUSD))
            .thenReturn(positionMock);

        assertThat(positionUtil.positionOrders(instrumentEURUSD), equalTo(positionMock));
    }
}
